use the expense class
